backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
psycopg2-binary
pytest
httpx
alembic
boto3
python-multipart
